Fix transaction data race.
We would occasionally see broken transactions and broken db connections after running for some time.

After much digging and experimenting, it turns out that this pathological case happens when a db connection is shared between goroutines, and that db connection is running a transaction on one as the other goroutine performs other queries.

Obviously, we shouldn't do that!

Change all places where we invoke a goroutine and pass along a data client to create a new data client (and thus, a new connection).

Most of these places are actually fine as they do not use the data client themselves after creating the goroutine, except for one code path:

startPhase -> checkBranch -> handleNewCommitsForBranch -> go StartTrain() -> ...

For simplicity, whenever we create a goroutine, we should create a new data client.